improve test by checking challenge number in campaign + competence AND move around files
