--- high-level-tests/e2e-playwright/next-challenge-regression/campaign-evaluation.json
@@ -1,1 +1,0 @@
-{"challengeImprints":["3256","697","21967","9531","21085","7480","14106","29915","17793","2160","19273","15062","9354","28533","10031","6152","10676","19598","6683","8175","17137","13121","590","2737","13396","30588","18854","1353","16073","12579","17957","2426","1680","10641","11735","25631","11697","22424","5847","11788","9789","21926","19922","5343","14412","6477","7699","5527","6023","3008","29013","12994","20723","19114","16190","13991","14753","22772","13296","13146","13752","2816","26972","22013","23506","13534","12582","19761","1234","4467","2417","2741","4609","14742","16941","2358","220","316","13338","22041","16078","1362","12410","23453","9400","5843","7147","684","13186","13266","24340","14731","17681","4170","12428","1779","14265","15741","1895","540","26963","14323","259","6167","5062","15228","13930","1924","12718","1583","2217","27992","4978","2869","19705","13878","12740","27656","11091","15426","6711","7987","18298","13013","18631","17394","544","14771","12082","975","10241","13053","11561","2803","7776","12860","13003","18050","12814","22580","29443","753","13041","1880","18160","9120","17034","6610","15198","13778","6448","350","6613","12466","4433","21942","23888","2156","8077","1628","12438","6579","10307","2227","29143","4293","21876","12890","13858","12852","4135","24484","4696","2017","6944","1348","10028","11658","12155","8311","11419","10327","4507","6202","1536","4110","3373","14286"],"masteryPercentages":["58","71","47","55","50","57","50","54","44","60","69","74","58","44","82","45"]}
