--- high-level-tests/e2e-playwright/next-challenge-regression/competence-evaluation.json
@@ -1,1 +1,0 @@
-{"challengeImprints":["13991","22682","11262","1897","1421","7184","46","15435","4708","12459","2149","9463","8064","3015","12994","24369","2360","12882","3439","28533","4609","9531","13121","14611","16941","13930","14441","12438","1362","22041","12560","7147","2426","12155","27231","11788","7480","13629","4642","13429","13310","13023","4354","9354","3479","9850","4135","5527","2869","9386","13212","12579","6202","2227","5415","4293","20277","26350","7857","24592","4679","26914","20173","23334","30397","14253","10968","5171","6996","16015","10534","2651","14592","16012","20902","9877","8495","10307","15385","11476","24447","8236","2585","5702","15909","6043","5787","29600","22972","22485","30451","6613","13534","7776","23761","25349","8175","27855","30607","25326","17137","24437","29040","29308","15062","30431","16336","9527","27653","30138","2044","9120","8311","13858","10740","30414"],"levels":["1","1","1","1","1"],"pixScores":["13","12","14","8","9"]}
