--- high-level-tests/e2e-playwright/tests/data/demo.json
@@ -1,1 +1,0 @@
-{"challengeImprints":["1489","17009","11969","16517","21171","9625","27231","1546","907","3214","16373","14820","14403","14053","18984","15525","7170","13983","9793","8220","11689","5950","6448","9980","21158","16858","8779","8336","15420","12463"]}
